fixes after base_agent move
[ghstack-poisoned]
