+++ craftassist/agent/dialogue_objects/get_memory_handler.py
@@ -16,7 +16,7 @@ from .spatial_reasoning import ComputeLocations
 from .point_target import PointTargetInterpreter
 from droidlet.base_util import ErrorWithResponse
 from ..tasks import Build, Point
-from droidlet.dialog.ttad.generation_dialogues import prepend_a_an
+from droidlet.dialog.ttad.generation_dialogues.generate_utils import prepend_a_an
 
 
 class MCGetMemoryHandler(GetMemoryHandler):
