+++ tools/data_scripts/fetch_aws_models.sh
@@ -36,18 +36,20 @@ cd $ROOTDIR
 echo "====== Downloading http://craftassist.s3-us-west-2.amazonaws.com/pubr/${MODELS_DIRNAME}_${CHECKSUM}.tar.gz to $ROOTDIR/${MODELS_DIRNAME}_${CHECKSUM}.tar.gz ======"
 curl http://craftassist.s3-us-west-2.amazonaws.com/pubr/${MODELS_DIRNAME}_${CHECKSUM}.tar.gz -o $MODELS_DIRNAME.tar.gz 
 
-if [ -d "${AGENT}/agent/models" ]
+if [ -d "agents/${AGENT}/models" ]
 then
 	echo "Overwriting models directory"
-	rm -rf $AGENT/agent/models/
+	rm -rf agents/${AGENT}/models
 fi
 
-mkdir -p $AGENT/agent/models/
+mkdir -p agents/${AGENT}/models
 
-tar -xzvf $MODELS_DIRNAME.tar.gz -C $AGENT/agent/models/ --strip-components 1 || echo "Failed to download and unarchive. Please make sure the file: ${MODELS_DIRNAME}_${CHECKSUM}.tar.gz exists on S3." 
+tar -xzvf $MODELS_DIRNAME.tar.gz -C agents/${AGENT}/models --strip-components 1 || echo "Failed to download and unarchive. Please make sure the file: ${MODELS_DIRNAME}_${CHECKSUM}.tar.gz exists on S3." 
 
 if [ $AGENT == "locobot" ]; then
     curl https://locobot-bucket.s3-us-west-2.amazonaws.com/perception_models.tar.gz -o locobot_models.tar.gz
-    tar -xzvf locobot_models.tar.gz -C $AGENT/agent/models/
-	curl https://locobot-bucket.s3-us-west-2.amazonaws.com/perception_test_assets.tar.gz | tar -xzv -C droidlet/perception/tests/test_assets/
+    tar -xzvf locobot_models.tar.gz -C agents/${AGENT}/models
+
+    mkdir -p droidlet/perception/tests/test_assets/
+    curl https://locobot-bucket.s3-us-west-2.amazonaws.com/perception_test_assets.tar.gz | tar -xzv -C droidlet/perception/tests/test_assets/
 fi
